feat(timesheet): implement export to IMS functionality
This commit introduces the functionality to export timesheet data to IMS.
It includes:

- A new module `export_to_ims.js` to handle the logic for toggling the
  table to IMS view, formatting dates, updating the table, navigation
  URLs, and table headers.
- Refactoring the `timesheet_overview.js` to import and use the new
  module, removing duplicated code.
- Adding a query parameter check to automatically toggle the IMS view
  if the `export_to_ims` parameter is present in the URL.
- The export to IMS button now triggers the toggleTableToIMS function
  when clicked.
